sunxi: H5/SPL: add 32-bit SPL config with stub DRAM init
Introduce a 32-bit SPL defconfig for Allwinner H5 boards. This is needed
to link it with the 32-bit libdram later.
In preparation for that separate the existing H3 DRAM build from the
shared H3 Kconfig symbol and introduce a separate DRAM setup file for
the H5, which is just an empty stub for now.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>