sunxi: introduce Allwinner H5 config option
The Allwinner H5 Soc is bascially an H3 with high SRAM and ARMv8 cores.
As the peripherals and the pinmuxing are almost identical, we piggy
back on the MACH_SUN8I_H3 config symbol.
Since the DRAM controller is unknown and not working with neither the
H3 nor A64 routines (out of the box, at least), we define it as a 64-bit
only non-SPL board for now and use boot0 as an SPL replacement.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>